// src/types/api.types.ts
export type Product = {
  id: number;
  title: string;
  price: number;
  description: string;
  images: string;
  created_at: string;
  updated_at: string;
  realPrice:number;
  weight:number
};
export type ProductsWithRealPrices = {
  id: number;
  title: string;
  price: number;
  description: string;
  images: string;
  realPrice:number;
  weight:number
};
// 1- Product Type Enum
export enum ProductType {
  GOLD_18 = 'gold_18',
  SAHAT_BAHAR = 'sahat_bahar',
  SAHAT_EMAMI = 'sahat_emami',
  NIM_SAHAT = 'nim_sahat',
  ROB_SAHAT = 'rob_sahat',
  SAHAT_SOT = 'sahat_sot',
}
export type AddProduct = {
  id: number;
  title: string;
  price: number;
  description: string;
  images: string[];
  created_at: string;
  updated_at: string;
  priceType:string ; 
  productType:ProductType;
  wages:number;
  profit:number;
  tax:number;
  weight:number;
  weightSot:number;
  gemPrice:number;
  gemActive:number;
  priceValue:number;
};